requirements.txt
fastapi
uvicorn
streamlit
faiss-cpu
sentence-transformers
pymupdf
python-multipart
requests
ollama
